requirements.txt
pyjwt

alabaster==0.7.13
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astroid==2.15.8
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Automat==0.8.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.13.0
azure-servicebus==7.11.0
azure-storage-blob==12.18.2
azure-storage-common==2.1.0
azure-storage-file-share==12.12.0
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
billiard==3.6.4.0
binaryornot==0.4.4
black==23.3.0
bleach==6.1.0
blinker==1.4
boto3==1.34.93
botocore==1.34.93
Brotli==1.1.0
build==0.10.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==23.4.4
Collectfast==2.2.0
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.5.0
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cryptography==40.0.2
cssbeautifier==1.15.1
cssselect2==0.7.0
dal-admin-filters==1.1.0
dbus-python==1.2.16
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
distlib==0.3.8
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.11
django-allauth==0.62.1
django-anymail==10.3
django-appconf==1.0.5
django-autocomplete-light==3.9.7
django-celery-beat==2.5.0
django-cid==2.3
django-ckeditor==6.5.1
django-compressor==4.3.1
django-cors-headers==4.3.1
django-coverage-plugin==3.1.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-js-asset==2.0.0
django-log-viewer==1.1.7
django-model-utils==4.5.0
django-redis==5.4.0
django-rest-swagger==2.2.0
django-simple-history==3.3.0
django-storages==1.14.2
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-timezone-field==5.0
django-utils-six==2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==3.14.5
djlint==1.34.1
dnspython==2.6.1
docutils==0.19
drf-spectacular==0.27.2
drf-yasg==1.21.5
drf-yasg2==1.19.4
EditorConfig==0.12.4
email_validator==2.1.1
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
factory-boy==3.3.0
Faker==24.11.0
fastapi==0.111.0
fastapi-cli==0.0.2
fastjsonschema==2.19.1
filelock==3.13.4
flake8==6.0.0
flake8-isort==6.0.0
flower==1.2.0
fonttools==4.51.0
fqdn==1.5.1
future==0.18.3
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
hiredis==2.3.2
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.14.0
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
hyperlink==19.0.0
identify==2.5.36
idna==3.4
imageio==2.34.1
imageio-ffmpeg==0.4.9
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
incremental==16.10.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.29.4
ipython==8.12.3
isodate==0.6.1
isoduration==20.11.0
isort==5.13.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.25
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
keyring==18.0.1
kombu==5.3.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.10.0
livereload==2.6.3
lxml==5.2.1
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==4.2.0
moviepy==1.0.3
msal==1.22.0
msal-extensions==1.0.0
mypy==1.7.1
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.4
nodeenv==1.8.0
notebook_shim==0.2.4
numpy==1.24.3
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
orjson==3.10.3
overrides==7.7.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pbr==6.0.0
pdfminer.six==20200517
pdfplumber==0.5.24
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.3.0
pip-tools==6.13.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pluggy==1.5.0
portalocker==2.7.0
pre-commit==3.3.1
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-c==3.1.18
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pydyf==0.9.0
pyflakes==3.0.1
Pygments==2.17.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.7.0
pylint==2.17.7
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==5.1.0
pyOpenSSL==19.0.0
pyphen==0.14.0
pypng==0.20220715.0
pyproject_hooks==1.0.0
pyrsistent==0.15.5
pyserial==3.4
pytest==8.2.0
pytest-django==4.8.0
pytest-sugar==1.0.0
python-apt==2.0.1+ubuntu0.20.4.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-docx==0.8.11
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
pyzmq==26.0.2
qrcode==7.4.2
rcssmin==1.1.1
redis==5.0.4
referencing==0.35.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rjsmin==1.2.1
rpds-py==0.18.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
ruff==0.4.2
s3transfer==0.10.1
SecretStorage==2.3.1
Send2Trash==1.8.3
service-identity==18.1.0
shellingham==1.5.4
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sos==4.5.6
soupsieve==2.5
SpeechRecognition==3.10.3
Sphinx==6.2.1
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
ssh-import-id==5.10
stack-data==0.6.3
starlette==0.37.2
stevedore==5.2.0
systemd-python==234
tablib==3.4.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.4
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
Twisted==18.9.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.11.0
tzdata==2023.3
uamqp==1.6.4
ubuntu-advantage-tools==8001
ufw==0.36
ujson==5.9.0
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
virtualenv==20.25.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wadllib==1.3.3
Wand==0.6.13
watchdog==4.0.0
watchfiles==0.19.0
wcwidth==0.2.13
weasyprint==61.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.2
xlwt==1.3.0
zipp==3.18.1
zope.interface==4.7.1
zopfli==0.2.3
bcrypt
passlib
